# Evaluation support for simple policies