--- a/go.mod
+++ b/go.mod
@@ -3,11 +3,10 @@ module github.com/google/cel-policy-templates-go
 go 1.12
 
 require (
-	github.com/golang/protobuf v1.4.0
-	github.com/google/cel-go v0.4.2
+	github.com/golang/protobuf v1.4.2
+	github.com/google/cel-go v0.5.0
 	github.com/kr/pretty v0.1.0 // indirect
-	google.golang.org/genproto v0.0.0-20200305110556-506484158171
+	google.golang.org/genproto v0.0.0-20200515170657-fc4c6c6a6587
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
-	gopkg.in/yaml.v2 v2.2.8 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20191010095647-fc94e3f71652
+	gopkg.in/yaml.v3 v3.0.0-20200506231410-2ff61e1afc86
 )