--- a/policy/testdata/invalid_instance_prop/instance.parse.out
+++ b/policy/testdata/invalid_instance_prop/instance.parse.out
@@ -1,0 +1,1 @@
+1~2~kinds: 3~"GrantTemplate"