… go.sum
… policy/compiler/compiler_test.go
… policy/engine.go
… policy/engine_test.go
… policy/evaluator.go
… policy/examples/allowed_resource_types/evaluator.yaml
… policy/examples/allowed_resource_types/instance.yaml
… policy/examples/required_labels/evaluator.yaml
… policy/examples/required_labels/instance.yaml
… policy/examples/required_labels/template.yaml
… policy/examples/resource_types/instance.yaml
… policy/examples/resource_types/template.yaml
… policy/examples/restricted_destinations/evaluator.yaml
… policy/examples/restricted_destinations/instance.yaml
… policy/examples/restricted_destinations/template.yaml
… policy/examples/sensitive_data/evaluator.yaml
… policy/examples/sensitive_data/instance.yaml
… policy/examples/sensitive_data/template.yaml
… policy/examples/timed_contract/instance.yaml
… policy/examples/timed_contract/template.yaml
… policy/file.go
… policy/instance.go
… policy/model/types.go
… policy/parser/parser.go
… policy/parser/yml/parser_test.go
… policy/runtime/evaluator.go
… policy/runtime/options.go
… policy/runtime/runtime.go
… policy/template.go
… policy/test/env.go