The following PR introduces several key features:
- Evaluation support for simple declarative policies
- CEL-based validation of policy instances
- Policy instance selector support
- Shared parse / compile / eval testdata

There are some improvements which can be made for surfacing multiple errors back from the `Engine#Eval` phase, and in checking schema content for the proper structure, e.g. if `properties` are defined the `type: object` field-value must be set. These will be addressed in future PRs.

Future PRs will support expressive policy instances with embedded CEL expressions.