@TristonianJones 2020-05-17
Note, the new evaluation stack has significant performance improvements over the prototype:

```
benchmark                                                                       old ns/op     new ns/op     delta
BenchmarkEnforcer/sensitive_data_prefix_same_location-8                         940           648           -31.06%
BenchmarkEnforcer/sensitive_data_prefix_diff_location-8                         3184          2493          -21.70%
BenchmarkEnforcer/sensitive_data_label_same_location-8                          962           651           -32.33%
BenchmarkEnforcer/sensitive_data_label_diff_location-8                          8048          5422          -32.63%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  9118          5352          -41.30%
BenchmarkEnforcer/timed_contract_valid-8                                        3393          1385          -59.18%
BenchmarkEnforcer/timed_contract_no_match-8                                     994           606           -39.03%
BenchmarkEnforcer/timed_contract_invalid-8                                      3443          1644          -52.25%
BenchmarkEnforcer/restricted_destinations_valid_location-8                      5237          4269          -18.48%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     3453          2567          -25.66%
BenchmarkEnforcer/restricted_destinations_restricted_location_ir_national-8     2247          1349          -39.96%
BenchmarkEnforcer/resource_types_denied_request-8                               4485          3204          -28.56%

benchmark                                                                       old allocs     new allocs     delta
BenchmarkEnforcer/sensitive_data_prefix_same_location-8                         7              2              -71.43%
BenchmarkEnforcer/sensitive_data_prefix_diff_location-8                         20             13             -35.00%
BenchmarkEnforcer/sensitive_data_label_same_location-8                          7              2              -71.43%
BenchmarkEnforcer/sensitive_data_label_diff_location-8                          51             32             -37.25%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  51             29             -43.14%
BenchmarkEnforcer/timed_contract_valid-8                                        12             5              -58.33%
BenchmarkEnforcer/timed_contract_no_match-8                                     8              3              -62.50%
BenchmarkEnforcer/timed_contract_invalid-8                                      14             9              -35.71%
BenchmarkEnforcer/restricted_destinations_valid_location-8                      38             18             -52.63%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     22             14             -36.36%
BenchmarkEnforcer/restricted_destinations_restricted_location_ir_national-8     12             7              -41.67%
BenchmarkEnforcer/resource_types_denied_request-8                               28             19             -32.14%

benchmark                                                                       old bytes     new bytes     delta
BenchmarkEnforcer/sensitive_data_prefix_same_location-8                         471           32            -93.21%
BenchmarkEnforcer/sensitive_data_prefix_diff_location-8                         769           224           -70.87%
BenchmarkEnforcer/sensitive_data_label_same_location-8                          471           32            -93.21%
BenchmarkEnforcer/sensitive_data_label_diff_location-8                          1566          697           -55.49%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  1575          640           -59.37%
BenchmarkEnforcer/timed_contract_valid-8                                        660           176           -73.33%
BenchmarkEnforcer/timed_contract_no_match-8                                     485           48            -90.10%
BenchmarkEnforcer/timed_contract_invalid-8                                      690           248           -64.06%
BenchmarkEnforcer/restricted_destinations_valid_location-8                      1101          336           -69.48%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     822           280           -65.94%
BenchmarkEnforcer/restricted_destinations_restricted_location_ir_national-8     624           160           -74.36%
BenchmarkEnforcer/resource_types_denied_request-8                               1662          1049          -36.88%
```